- Added Parallel port to Visual C++. - Applied __SCREEN_LINES_LIMIT to VGA timing again. - Implemented parallel port fully. - Fixed mouse timing initializing. - Fixed keyboard timing initializing. - Fixed Adlib note on type warning. - Implemented missing Parallel port data buffer for reading it back. - Implemented UART ports only when used. - Implemented VGA Vertical Retrace Interrrupt and related flags handling according to documentation. - Decreased VGA Sequencer pixel block size for better CPU response. - Improved VGA Vertical Retrace End Register documentation.

+++ hardware/adlib.c
@@ -211,7 +211,7 @@ byte outadlib (uint16_t portnum, uint8_t value) {
 		{ //octave, freq, key on
 			if ((portnum & 0xF) > 8) goto unsupporteditem; //Ignore A9-AF!
 			portnum &= 0xF; //Only take the lower nibble (the channel)!
-			writeadlibKeyON(portnum,0); //Write to this port! Don't force the key on!
+			writeadlibKeyON((byte)portnum,0); //Write to this port! Don't force the key on!
 		}
 		else if (portnum == 0xBD) //Percussion settings etc.
 		{
